+++ lib/src/align.dart
@@ -1,0 +1,1 @@
+enum Align { stretch, start, end, center }