+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+- Move property hAlign and vAlign from PanelComponent to Component.
+- Add textAlign to Component.
+
 ## 0.0.1
 
 - Init version